MAINT: Renamed modules to reference basemap, updated unitests to reflect new names.
